+++ src/main/resources/database/postgresql/1/patch_22.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS ember_schema.event_layout_field;
+DROP TABLE IF EXISTS ember_schema.event_layout;